+++ lib/active_model_serializers/adapter/json_api/pagination_links.rb
@@ -12,8 +12,9 @@ module ActiveModelSerializers
         end
 
         def serializable_hash(options = {})
+          per_page = collection.try(:per_page) || collection.try(:limit_value) || collection.size
           pages_from.each_with_object({}) do |(key, value), hash|
-            params = query_parameters.merge(page: { number: value, size: collection.size }).to_query
+            params = query_parameters.merge(page: { number: value, size: per_page }).to_query
 
             hash[key] = "#{url(options)}?#{params}"
           end